[IMP] Simpler get_files_recursive
The differences with this function compared to the previous version is that
the files from hidden directories (starting with a `.`) will not be listed.

Usually, you do not want your files inside the dotted directories to be
modified by a script. This is a more modern approach available since 3.4.

Also, make it a generator to have instantaneous feedback.

It is almost 2 times faster directly if you want to obtain a list.